Controller refactor for subscriber route (#181)
