+++ routes/adminVerification.js
@@ -1,3 +1,5 @@
+/* BASE PATH: /adminVerify */
+
 const express = require("express");
 const router = express.Router();
 const AdminVerification = require("../controllers/adminVerification.controller.js");
