+++ routes/googleLogin.js
@@ -1,3 +1,5 @@
+/* BASE PATH: /googleLogin */
+
 const express = require("express");
 const router = express.Router();
 const GoogleLogin = require("../controllers/googleLogin.controller.js");
